indirect_kalman_6dof: tuned based on bench bag data, fixed bug where inverse pose orientation was published.
As of now it is looking "ok" in subsim. It would appear to lead the
legacy_ukf during fast accelerations, something I was hoping we might
see due to not stretching DVL updates across pings. However, the
accelerometer hack is perceptible and probably is causing errors,
though without it we do begin to drift on pitch and roll and lose
accel biases while stationary. I suspect the reason we don't see the
legacy_ukf's equivalent hack is because it averages accelerometer
readings, which I dislike conceptually but might be worth a try. A
better option might be a 2nd order lowpass filter on all sensor
readings, so that delay is mostly constant (and small) across all the
sensors.

One strange behavior I can't yet explain is that pushing the sub down
rapidly causes this filter to yaw, which would suggest a math error
somewhere in the DVL or depth sensor update. Shoving it around in the
sim can get other odd behaviors to occur, so I'm probably going to
need to recreate this in matlab where I can see the effects more
closely.