indirect_kalman_6dof: Removed accelerometer bias estimation.
The accel biases were never really converging in pool testing. My
theory is that our orientation error is high enough such that the
biases are never really observable. Removing the accel biases also
means that the DVL is sufficient to observe roll and pitch error, so
we can eliminate the accelerometer hack and actually gain some
orientation accuracy and stability.