indirect_kalman_6dof: Added a service to set position of filter.
Will be used for barrel rolls to clean up the stabilization at the end
of the roll. Ignores the z component, since this is observed so changing it
from under the filter just makes it angry. It also zeros velocity, since
in the barrel roll case the velocity is pretty wrong and until we get a DVL
ping or two the sub will move to one side. This service is also handy for
reseting the sub to the origin. Just kill the sub, use rqt_gui or something
to invoke with (0, 0, 0), and the sub is now at the origin.