c3_trajectory_generator: only use time from one source instead of mixing them
This (likely) fixes a lot of weirdness that I've seen coming from
c3_trajectory_generator, which primarily manifests as pauses and stutters
after unkills in the outgoing trajectory. This is caused by using timestamps
from both the /odom topic and the local machine and not really tolerating any
significant difference between them.